tests/submission/PhilipTaylor/canvas/size.attributes.idl.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.idl</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.idl</h1>
<p class="desc">Getting/setting width/height IDL attributes</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.width">size.width</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.size.height">size.height</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Getting/setting width/height IDL attributes");
_addTest(function(canvas, ctx) {

canvas.width = "100";
canvas.height = "100";
_assertSame(canvas.width, 100, "canvas.width", "100");
_assertSame(canvas.height, 100, "canvas.height", "100");

canvas.width = "+1.5e2";
canvas.height = "0x96";
_assertSame(canvas.width, 150, "canvas.width", "150");
_assertSame(canvas.height, 150, "canvas.height", "150");

canvas.width = 200 - Math.pow(2, 32);
canvas.height = 200 - Math.pow(2, 32);
_assertSame(canvas.width, 200, "canvas.width", "200");
_assertSame(canvas.height, 200, "canvas.height", "200");

canvas.width = 301.999;
canvas.height = 301.001;
_assertSame(canvas.width, 301, "canvas.width", "301");
_assertSame(canvas.height, 301, "canvas.height", "301");

canvas.width = "400x";
canvas.height = "foo";
_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 0, "canvas.height", "0");


});
</script>

<div id="log"></div>
</body></html>
